Machine-Learning-Damaged-Box-Classifier/requirements.txt
streamlit
numpy
tensorflow
opencv-python
pillow